[Refactor] Extract common format scanner context types
