+++ be/module_boundary_manifest.json
@@ -319,6 +319,7 @@
         "be/src/formats/disk_range.hpp",
         "be/src/formats/file_writer.*",
         "be/src/formats/io/**",
+        "be/src/formats/scan_context.h",
         "be/src/formats/utils.h"
       ],
       "allowed_include_prefixes": [
@@ -327,6 +328,7 @@
         "formats/disk_range.hpp",
         "formats/file_writer.h",
         "formats/io/",
+        "formats/scan_context.h",
         "formats/utils.h",
         "exprs/",
         "runtime/",
