+++ be/src/formats/parquet/meta_helper.h
@@ -48,7 +48,7 @@ public:
             : _file_metadata(file_metadata), _case_sensitive(case_sensitive) {}
     virtual ~MetaHelper() = default;
 
-    virtual void prepare_read_columns(const std::vector<HdfsScannerContext::ColumnInfo>& materialized_columns,
+    virtual void prepare_read_columns(const std::vector<FormatColumnInfo>& materialized_columns,
                                       const std::vector<ColumnAccessPathPtr>* column_access_paths,
                                       std::vector<GroupReaderParam::Column>& read_cols,
                                       std::unordered_set<std::string>& existed_column_names) const = 0;
@@ -76,7 +76,7 @@ public:
             : MetaHelper(file_metadata, case_sensitive) {}
     ~ParquetMetaHelper() override = default;
 
-    void prepare_read_columns(const std::vector<HdfsScannerContext::ColumnInfo>& materialized_columns,
+    void prepare_read_columns(const std::vector<FormatColumnInfo>& materialized_columns,
                               const std::vector<ColumnAccessPathPtr>* column_access_paths,
                               std::vector<GroupReaderParam::Column>& read_cols,
                               std::unordered_set<std::string>& existed_column_names) const override;
@@ -96,7 +96,7 @@ public:
 
     ~LakeMetaHelper() override = default;
 
-    void prepare_read_columns(const std::vector<HdfsScannerContext::ColumnInfo>& materialized_columns,
+    void prepare_read_columns(const std::vector<FormatColumnInfo>& materialized_columns,
                               const std::vector<ColumnAccessPathPtr>* column_access_paths,
                               std::vector<GroupReaderParam::Column>& read_cols,
                               std::unordered_set<std::string>& existed_column_names) const override;
