+++ be/test/exec/hdfs_scanner/jni_scanner_test.cpp
@@ -63,7 +63,7 @@ public:
         const auto& slots = tuple_desc->slots();
         for (int i = 0; i < slots.size(); i++) {
             SlotDescriptor* slot = slots[i];
-            HdfsScannerContext::ColumnInfo info;
+            FormatColumnInfo info;
             info.idx_in_chunk = i;
             info.slot_desc = slot;
             ctx->materialized_columns.push_back(info);
