+++ be/test/formats/parquet/column_converter_test.cpp
@@ -30,7 +30,7 @@
 
 namespace starrocks::parquet {
 
-static HdfsScannerStats g_hdfs_stats{};
+static FormatScannerStats g_hdfs_stats{};
 
 class ColumnConverterTest : public testing::Test {
 public:
