+++ be/test/formats/parquet/page_index_test.cpp
@@ -34,7 +34,7 @@
 
 namespace starrocks::parquet {
 
-static HdfsScannerStats g_hdfs_stats;
+static FormatScannerStats g_hdfs_stats;
 using starrocks::HdfsScannerContext;
 
 class PageIndexTest : public testing::Test {
