+++ be/test/formats/parquet/page_reader_test.cpp
@@ -35,7 +35,7 @@ public:
 
 TEST_F(ParquetPageReaderTest, Normal) {
     std::string buffer;
-    HdfsScannerStats stats;
+    FormatScannerStats stats;
 
     // page 0
     {
@@ -101,7 +101,7 @@ TEST_F(ParquetPageReaderTest, Normal) {
 
 TEST_F(ParquetPageReaderTest, ExtraBytes) {
     std::string buffer;
-    HdfsScannerStats stats;
+    FormatScannerStats stats;
 
     // page 0
     {
