+++ be/test/formats/parquet/parquet_test_util/util.h
@@ -65,12 +65,11 @@ public:
         return row_desc->tuple_descriptors()[0];
     }
 
-    static void make_column_info_vector(const TupleDescriptor* tuple_desc,
-                                        std::vector<HdfsScannerContext::ColumnInfo>* columns) {
+    static void make_column_info_vector(const TupleDescriptor* tuple_desc, std::vector<FormatColumnInfo>* columns) {
         columns->clear();
         for (int i = 0; i < tuple_desc->slots().size(); i++) {
             SlotDescriptor* slot = tuple_desc->slots()[i];
-            HdfsScannerContext::ColumnInfo c;
+            FormatColumnInfo c;
             c.idx_in_chunk = i;
             c.slot_desc = slot;
             columns->emplace_back(c);
